# contracts/browser-capture-reader/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fragments Engine Browser Capture and Reader API",
    "version": "1.0.0",
    "description": "Canonical FE contract for manifest-first browser capture, media custody handoff, capability discovery, and Reader projections. Contract presence is distinct from runtime operation readiness reported by GET /v1/capabilities."
  },
...
  "servers": [
    {"url": "http://127.0.0.1:8091"}
  ],
  "paths": {
...
    "/v1/captures": {
      "post": {
        "operationId": "createCapture",
        "summary": "Atomically accept a capture manifest before optional asset transfer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "./schema/capture-envelope.schema.json"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "New capture attempt accepted",
            "content": {"application/json": {"schema": {"$ref": "./schema/capture-response.schema.json"}}}
          },
          "200": {
            "description": "Idempotent replay of an accepted capture ID",
            "content": {"application/json": {"schema": {"$ref": "./schema/capture-response.schema.json"}}}
          },
          "400": {"$ref": "#/components/responses/Problem"},
          "409": {"$ref": "#/components/responses/Problem"},
          "413": {"$ref": "#/components/responses/Problem"}
        }
      }
    },
    "/v1/captures/{captureId}": {
      "get": {
        "operationId": "getCapture",
        "summary": "Get accepted manifest and per-asset client contribution status",
        "parameters": [
          {"$ref": "#/components/parameters/CaptureID"}
        ],
        "responses": {
          "200": {
            "description": "Capture status",
            "content": {"application/json": {"schema": {"$ref": "./schema/capture-status.schema.json"}}}
          },
          "404": {"$ref": "#/components/responses/Problem"}
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "CaptureID": {
        "name": "captureId",
        "in": "path",
        "required": true,
        "schema": {"type": "string", "minLength": 1}
      },
...
    },
    "responses": {
      "Problem": {
        "description": "Standard FE problem response",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "./schema/error.schema.json"}
          }
        }
      }
    }
  }
}